% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts_enrollment}
\alias{districts_enrollment}
\title{NYS District Enrollment}
\format{
a data frame with enrollment information about each district in NYS
}
\source{
http://www.p12.nysed.gov/irs/statistics/enroll-n-staff/
}
\description{
Enrollment data for school districts in NYS.
}
\keyword{datasets}
