% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAndSaveRDS.R
\name{downloadAndSaveRDS}
\alias{downloadAndSaveRDS}
\title{Download a ZIP file, extract, and save the data as an RDS file.}
\usage{
downloadAndSaveRDS(
  url,
  filename = tools::file_path_sans_ext(basename(url)),
  download.dir = "data-raw/",
  out.dir = "data/",
  filetypes = c("mdb", "aacdb", "csv"),
  redownload = FALSE,
  cleanup = FALSE
)
}
\arguments{
\item{url}{the url of the ZIP file containing the data.}

\item{filename}{the name of the RDS file to write (excluding file extension).}

\item{download.dir}{directory to download the ZIP file.}

\item{out.dir}{directory to write the RDS file.}

\item{filetypes}{file extensions to look for in the ZIP file to convert to RDS.}

\item{redownload}{whether to redownload the ZIP file.}

\item{cleanup}{delete the downloaded ZIP file if the RDS file was created successfully.}
}
\value{
the data object read. This would be a data frame (for CSV files) or
a list (for MDB and AACDB files).
}
\description{
This is a utility function that will download a zip file and extract the
data.
}
