% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schools_enrollment}
\alias{schools_enrollment}
\title{NYS School Enrollment}
\format{
a data frame with enrollment information about each school in NYS
}
\source{
http://www.p12.nysed.gov/irs/statistics/enroll-n-staff/
}
\description{
Enrollment data for schools in NYS.
}
\keyword{datasets}
